Phase 3 (P3.64): extract Canvas getContext into CanvasBinding; close last element-member residue
The canvas.getContext("2d") binding was the last element-member callback still
living in the mixed DomBridge/JsObjects.cs file. It could not be peeled out
earlier because the 2D context was tied to the standalone
Broiler.HtmlBridge.Rendering project's CanvasCommandRecorder; Phase 6 (P6.1)
internalized the context into Dom and Phase 8 F1 (P8.9) dissolved Rendering into
Dom, which removes that coupling.

Extract getContext, BuildCanvas2DContext, and the sixteen JsUtilities…034…058Core
drawing callbacks (setFillStyle..restore, measureText) verbatim into the
co-located Dom.Features.CanvasBinding module, reached through the neutral static
DomBridge.TryGetAttribute — the binding needs no host contract. The call site
becomes Dom.Features.CanvasBinding.Install(obj, element), matching the other
element bindings; the #if BROILER_CLI guard is preserved. Behaviour-preserving,
no public-API change; CanvasContextBindingTests and the Module/JsObject/
ElementBinding/DomBinding suites (269 tests) pass.

Housekeeping to re-close the Phase 3 750-line ratchet and drop a stale Phase 6
guard:
- Split PositionArea.cs (had grown to 769 lines on main) — move the scroll-
  container geometry helpers into the sibling partial PositionArea.ScrollGeometry.cs,
  dropping the main file to 663 lines.
- Remove the stale Bridge_Rendering_Project_Graph_Does_Not_Reach_Html_Image guard
  and the dead Rendering entry in the file-size sweep: the Rendering project was
  deleted (P8.9), and Bridge_Dom_Project_Graph_Does_Not_Reach_Html_Image now
  covers the whole bridge-Dom graph.

Update the roadmap docs (overview status table, Phase 0–5 delivery log, remaining-
phases status) to record P3.64 and reflect that Phases 6/7/8 are complete.
