+++ docs/roadmap/htmlbridge-complexity-reduction-roadmap.md
@@ -262,7 +262,7 @@ the specifics per slice.
 | 0 — stabilize the boundary / baseline | Baseline established | Recorded in [Phase 0 baseline](htmlbridge-phase0-baseline.md); no explicit completion assertion. |
 | 1 — repair the project graph | **Complete** | None — all five work items landed. |
 | 2 — document services & single state authority | **Complete** | All work landed. The process-static per-element runtime tables are **fully de-globalized** to per-bridge instances — `ElementRuntimeState`/`GetElementRuntimeState` no longer exist (the last concern, inline-style, was moved and the type renamed `InlineStyleRuntimeState`), `PositionAreaResolutions` is a per-bridge instance table, and a boundary guard enforces "no process-static per-element table". The **"two simultaneous sessions are isolated"** exit criterion — previously believed blocked below the bridge — **is now met and verified (2026-07-21)**: the Broiler.JS engine already isolates two live `JSContext` instances (each `Eval` enters its own realm scope; the current context is `[ThreadStatic]` + `AsyncLocal`-scoped, not a last-wins global — proven by interleaved, concurrent-2000×, and stored-callback tests), so with the bridge state de-globalized, two simultaneous bridge sessions do not see each other's nodes/globals/mutations (`DomBridgeSessionLifetimeTests.Two_Simultaneous_Sessions_Do_Not_See_Each_Others_State`). Engine guard `patches/0009` is now **applied upstream** (pinned `Broiler.JS` `3f0c7054`). |
-| 3 — feature modules | **Complete (in-scope)** | All four exit criteria are met (verified 2026-07-21): ~60 feature-binding modules extracted with co-located tests and declared host contracts; the `DomBridge.cs` facade is **within** the 500–800-line target (701 lines) — the earlier "still to come" list (Element/geometry — P3.51, SVG — P3.50, Window/Document, Canvas) is done except **Canvas**, which is Phase-6-gated (its binding is tied to the Rendering project); and the 750-line file-size ratchet is fully closed (debt list empty, guard `HtmlBridgeArchitectureGuardTests`). Work item 6 — externalize the embedded content-rendering polyfill JS as a versioned asset — is done (P3.53). The only residue is the Phase-6-gated Canvas `getContext`/2D callbacks. |
+| 3 — feature modules | **Complete** | All four exit criteria are met (verified 2026-07-21): ~60 feature-binding modules extracted with co-located tests and declared host contracts; the `DomBridge.cs` facade is **within** the 500–800-line target (701 lines) — the earlier "still to come" list (Element/geometry — P3.51, SVG — P3.50, Window/Document, Canvas) is done; and the 750-line file-size ratchet is fully closed (debt list empty, guard `HtmlBridgeArchitectureGuardTests`). Work item 6 — externalize the embedded content-rendering polyfill JS as a versioned asset — is done (P3.53). **The last residue — the Phase-6-gated Canvas `getContext`/2D callbacks — is now closed (P3.64, 2026-07-23):** with Phase 6 / Phase 8 F1 (P8.9) dissolving `Broiler.HtmlBridge.Rendering` into `Dom`, the `CanvasCommandRecorder` coupling that pinned the callback in the mixed `JsObjects.cs` file is gone, so `getContext` + `BuildCanvas2DContext` + the `JsUtilities…034…058Core` drawing callbacks were extracted into the co-located `Dom.Features.CanvasBinding` module (`CanvasContextBindingTests` green). Every element-member callback is now a co-located feature module. |
 | 4 — eliminate parallel DOM state | **Complete (in-scope)** | The in-scope work landed (verified 2026-07-21): sentinel `#document`/`#fragment`/`#doctype`/`#subdoc-root` elements replaced by canonical types (item 1); the parallel `InnerHtml` string removed (item 3); the canonical-reuse promotions done (item 4 — P4.8–4.12, P4.16–4.19; `ChildIndexOf`→`IndexOfReference`); and **item 5 complete** — `CloneDomElement` delegates to canonical `DomNode.CloneNode` (P4.20) and `NormalizeNode` to canonical `DomNode.Normalize()` (P4.21, closing the P4.15 side-effect-coupling blocker via the mutation-notification consolidation onto `DomDocument.Mutated`). Item 5 was the **last in-scope item**. Residue: item 2's *full* ~200-site inline-style dict elimination (its baked-style split is done — P4.14; the remaining rewrite is **Phase-5-entangled** via the anchor resolver, so it's sequenced with Phase 5), and the full-corpus WPT/Acid CI gate (external). |
 | 5 — used-value behaviour into Layout | **In-scope terminal** | Anchor-track deletion complete through step 6 (verified 2026-07-21). Feature (b) visual-viewport/zoom: read (CSSOM) **and** render (pixel) sides validated on the engine used-value model — `NativeVisualViewport` is on, the zoom-serialization revert machinery is deleted, and `NativeZoom`/engine-render reftests pass; deleting the zoom bake is now only a *deployment* gate (enable `NativeZoom` at the external `CaptureService` renderer). Feature (a) dialog: native modal centering wired, modal box-chrome bake deleted, **and the native
 `::backdrop` box + top-layer paint cutover landed (2026-07-21)** — those submodule pieces
