+++ src/Broiler.HtmlBridge.Dom/DomBridge/Utilities.cs
@@ -489,91 +489,8 @@ public sealed partial class DomBridge
     // (Broiler.HtmlBridge.Dom.Features).
     // style / CSSStyleDeclaration (element.style, rule.style, getComputedStyle result) moved to the
     // Phase 3 (P3.14) StyleDeclarationBinding feature module (Broiler.HtmlBridge.Dom.Features).
-
-#if !BROILER_CLI
-    /// <summary>
-    /// Builds a minimal Canvas 2D rendering context exposing basic drawing
-    /// operations as defined in the HTML Canvas 2D Context specification.
-    /// Drawing commands are recorded but not rasterised in the current implementation.
-    /// </summary>
-    private static JSObject BuildCanvas2DContext(DomElement canvas)
-    {
-        var ctx = new JSObject();
-        int width = 300, height = 150;
-        if (TryGetAttribute(canvas, "width", out var w) && int.TryParse(w, out var pw)) width = pw;
-        if (TryGetAttribute(canvas, "height", out var h) && int.TryParse(h, out var ph)) height = ph;
-
-        var context2d = new CanvasRenderingContext2D(width, height);
-
-        // fillStyle (get/set)
-        ctx.FastAddProperty((KeyString)"fillStyle",
-            new JSFunction((in _) => new JSString(context2d.FillStyle), "get fillStyle"),
-            new JSFunction((in a) => JsUtilitiesSetFillStyle034Core(context2d, in a), "set fillStyle"),
-            JSPropertyAttributes.EnumerableConfigurableProperty);
-
-        // strokeStyle (get/set)
-        ctx.FastAddProperty((KeyString)"strokeStyle",
-            new JSFunction((in _) => new JSString(context2d.StrokeStyle), "get strokeStyle"),
-            new JSFunction((in a) => JsUtilitiesSetStrokeStyle036Core(context2d, in a), "set strokeStyle"),
-            JSPropertyAttributes.EnumerableConfigurableProperty);
-
-        // lineWidth (get/set)
-        ctx.FastAddProperty((KeyString)"lineWidth",
-            new JSFunction((in _) => new JSNumber(context2d.LineWidth), "get lineWidth"),
-            new JSFunction((in a) => JsUtilitiesSetLineWidth038Core(context2d, in a), "set lineWidth"),
-            JSPropertyAttributes.EnumerableConfigurableProperty);
-
-        // font (get/set)
-        ctx.FastAddProperty((KeyString)"font",
-            new JSFunction((in _) => new JSString(context2d.Font), "get font"),
-            new JSFunction((in a) => JsUtilitiesSetFont040Core(context2d, in a), "set font"),
-            JSPropertyAttributes.EnumerableConfigurableProperty);
-
-        // globalAlpha (get/set)
-        ctx.FastAddProperty((KeyString)"globalAlpha",
-            new JSFunction((in _) => new JSNumber(context2d.GlobalAlpha), "get globalAlpha"),
-            new JSFunction((in a) => JsUtilitiesSetGlobalAlpha042Core(context2d, in a), "set globalAlpha"),
-            JSPropertyAttributes.EnumerableConfigurableProperty);
-
-        // canvas property
-        ctx.FastAddProperty((KeyString)"canvas",
-            new JSFunction((in _) => new JSObject(), "get canvas"),
-            null, JSPropertyAttributes.EnumerableConfigurableProperty);
-
-        // Drawing methods
-        ctx.FastAddValue((KeyString)"fillRect", new JSFunction((in a) => JsUtilitiesFillRect044Core(context2d, in a), "fillRect", 4), JSPropertyAttributes.EnumerableConfigurableValue);
-
-        ctx.FastAddValue((KeyString)"strokeRect", new JSFunction((in a) => JsUtilitiesStrokeRect045Core(context2d, in a), "strokeRect", 4), JSPropertyAttributes.EnumerableConfigurableValue);
-
-        ctx.FastAddValue((KeyString)"clearRect", new JSFunction((in a) => JsUtilitiesClearRect046Core(context2d, in a), "clearRect", 4), JSPropertyAttributes.EnumerableConfigurableValue);
-
-        ctx.FastAddValue((KeyString)"beginPath", new JSFunction((in _) => JsUtilitiesBeginPath047Core(context2d, in _), "beginPath", 0), JSPropertyAttributes.EnumerableConfigurableValue);
-
-        ctx.FastAddValue((KeyString)"moveTo", new JSFunction((in a) => JsUtilitiesMoveTo048Core(context2d, in a), "moveTo", 2), JSPropertyAttributes.EnumerableConfigurableValue);
-
-        ctx.FastAddValue((KeyString)"lineTo", new JSFunction((in a) => JsUtilitiesLineTo049Core(context2d, in a), "lineTo", 2), JSPropertyAttributes.EnumerableConfigurableValue);
-
-        ctx.FastAddValue((KeyString)"arc", new JSFunction((in a) => JsUtilitiesArc050Core(context2d, in a), "arc", 5), JSPropertyAttributes.EnumerableConfigurableValue);
-
-        ctx.FastAddValue((KeyString)"closePath", new JSFunction((in _) => JsUtilitiesClosePath051Core(context2d, in _), "closePath", 0), JSPropertyAttributes.EnumerableConfigurableValue);
-
-        ctx.FastAddValue((KeyString)"fill", new JSFunction((in _) => JsUtilitiesFill052Core(context2d, in _), "fill", 0), JSPropertyAttributes.EnumerableConfigurableValue);
-
-        ctx.FastAddValue((KeyString)"stroke", new JSFunction((in _) => JsUtilitiesStroke053Core(context2d, in _), "stroke", 0), JSPropertyAttributes.EnumerableConfigurableValue);
-
-        ctx.FastAddValue((KeyString)"fillText", new JSFunction((in a) => JsUtilitiesFillText054Core(context2d, in a), "fillText", 3), JSPropertyAttributes.EnumerableConfigurableValue);
-
-        ctx.FastAddValue((KeyString)"strokeText", new JSFunction((in a) => JsUtilitiesStrokeText055Core(context2d, in a), "strokeText", 3), JSPropertyAttributes.EnumerableConfigurableValue);
-
-        ctx.FastAddValue((KeyString)"save", new JSFunction((in _) => JsUtilitiesSave056Core(context2d, in _), "save", 0), JSPropertyAttributes.EnumerableConfigurableValue);
-
-        ctx.FastAddValue((KeyString)"restore", new JSFunction((in _) => JsUtilitiesRestore057Core(context2d, in _), "restore", 0), JSPropertyAttributes.EnumerableConfigurableValue);
-
-        // measureText(text) — returns { width: ... }
-        ctx.FastAddValue((KeyString)"measureText", new JSFunction((in a) => JsUtilitiesMeasureText058Core(in a), "measureText", 1), JSPropertyAttributes.EnumerableConfigurableValue);
-
-        return ctx;
-    }
-#endif
+    // canvas.getContext("2d") + BuildCanvas2DContext moved to the Phase 3 (P3.64) CanvasBinding feature
+    // module (Broiler.HtmlBridge.Dom.Features), unblocked once Phase 6/P8.9 dissolved
+    // Broiler.HtmlBridge.Rendering into Dom.
 
 }
